Refactor patternDiagramLinearWithLanes to work differently
Becuase I have a bunch of diagrams that are all a bit different, it turns out
that it's better to define helpers locally in the function.  How you compute
the label, whether you use the whole or the part of an event, etc., is all
different enough between diagrams that it's better to handle that explicitly
per diagram instead of having helpers that have to handle all the different
kinds of diagrams.