Add the .diagram class for ghcisession blocks
This adds a second lua script that extracts the list of diagram blocks and
generates two files: a makefile with a target that lists the diagrams to build,
and a haskell file that contains the expression corresponding to the diagram.

With this scheme we don't have to maintain a duplicate list of diagrams that
appear in the document.  We also don't have to write the expression once, so
there's no danger of two copies of the expression string diverging.

We have to create a separate target in the makefile for the new
automatically-generated haskell file, since it'll complain about
missing top-level type signatures.