Add ghcisession for basicpattern.
Can't do the diagram this way since it gets the type
 $ ghci -e 'import Sound.Tidal.Context' -e ':t parseBP_E "a b c"'
 parseBP_E "a b c" :: (Enumerable a, Parseable a) => Pattern a

which is ambiguous in PatternExpressions.hs.

In this case, for now at least, I'll have to stick to the previous method of
setting up the diagram where I manually copy the expression string over to the
haskell diagram code.